Add Day 3 Advent of Code solution and update input encoding
Added a new solution for the 3rd day of Advent of Code. This solution includes the structure of 'Cord' to find neighbors and check the validity of the inputs. Also, the tasks encoding in the 'build.gradle.kts' file has been updated from 'utf-8' to 'UTF-8' to ensure consistency in encoding throughout the application environment. Test input and output files have also been created for testing purposes.